refector: CircleTitle , CategoryItem map사용
