+++ src/common/DrawingHeader /index.tsx
@@ -1,6 +1,8 @@
 import { styled } from 'styled-components';
 import Exit from '@/assets/Exit.png';
-import Circle from '@/components/DrawingResultCircle';
+
+const CircleText = ['우', '리', '의', '작', '품'];
+
 const DrawingHeader = () => {
   return (
     <PageHeader>
@@ -10,7 +12,9 @@ const DrawingHeader = () => {
           <Text>돌아가기</Text>
         </ExitBtn>{' '}
       </WrapBtn>
-      <Circle />
+      {CircleText.map((text, index) => (
+        <CircleWrap key={index}>{text}</CircleWrap>
+      ))}
     </PageHeader>
   );
 };
@@ -50,4 +54,18 @@ const Text = styled.span`
   font-size: 3rem;
   font-weight: bold;
   color: #000000;
+`;
+
+const CircleWrap = styled.div`
+  width: 8rem;
+  height: 8rem;
+  margin-left: 3rem;
+  background-color: #fff;
+  border: 0 solid;
+  border-radius: 50%;
+  text-align: center;
+  line-height: 8.5rem;
+  font-size: 5rem;
+  text-shadow: 1px 1px 2px #b8b8b8;
+  box-shadow: 4px 4px 3px -1px #a6a6a6;
 `;
